(BUGFIX) Replace validate_re() with assert_type()
As of the newest version of puppetlabs-stdlib validate_re and many other similar functions have been fully removed from the codebase after having been deprecated for some time.
As such it is necessary to use the built in assert_type command in its place.
Have chosen to use `Stdlib::Unixpath` over `Stdlib::Absolutepath` as this module does not support windows.